Committed as this is the exact code used to produce the files "6ss-256,90,32bins-xxxx-seed42-testwindivnodes.txt"
It checks out with the branch version of the code, revision 60, but... there are rounding errors. The rounding errors are entirely from (a) adding up the total regret with a down counting loop (branch) vs an up counting loop (trunk) and (b) calculating the last stratt via one minus the other stratt's (branch) vs the last regret over totalregret (trunk). The differences are huge after 1.6M iterations. (+-10% probabilities). I spent too long figuring out that those are the two places the rounding errors occur. The rounding errors are gone when you change the trunk code to have a down-counting loop and the branch code to calculate the final stratt the way it should be.

- fixed minor typo/bug in savestrategy.cpp

git-svn-id: file:///home/scott/regretitron_publish/regretitron_filtered3_svn/trunk@61 56bb13ec-af9c-4a7e-874f-62681e7d7c12