Committing some code that's been floating around uncommitted from last summer. I want to commit it because I've been using it lately and it works. I think it is IDENTICAL to the old code in 2 cases: (1) singlethreaded, and (2) multithreaded and using history bins.
All the changes are to using non-history bins. The changes are to set this case to work the same way as the history bins does, which only uses 1 lock, 1 signaler, and a bunch of boolean flags, instead of a lock for every bin on every round. The code in this case is UNTESTED. Looks decent, and might work though, but needs to be tested against single-threaded, and maybe even tested against the old revision.

git-svn-id: file:///home/scott/regretitron_publish/regretitron_filtered3_svn/trunk@87 56bb13ec-af9c-4a7e-874f-62681e7d7c12