Added the fabled nested bin support.
Easier than expected. I adapted the turn binning method to also nest flop bins, and
the river binning method to also nest turn bins. The process to create NxM nested bins,
is to first create N regular bins that are sorted by HSS, then for each of the N bins,
to further sort into M bins based on EV. Finally, the two bin numbers must be combined. I
correctly used the N-number as the major and the M-number as the minor in the combine.
Clearly the process for creating flop nested bins is the same as creating turn regular
bins, except I don't iterate through the possible turn cards.


git-svn-id: file:///home/scott/regretitron_publish/regretitron_filtered3_svn/trunk@158 56bb13ec-af9c-4a7e-874f-62681e7d7c12